gameStateHelpers WIP
